% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corls_test.R
\name{.getPermuteMatrix}
\alias{.getPermuteMatrix}
\title{Generate permutation matrix according to a schema.}
\usage{
.getPermuteMatrix(permutations, n, strata = NULL)
}
\arguments{
\item{permutations}{a list of control values for the permutations as returned
by the function \code{\link[permute]{how}}, or the number of
permutations required.}

\item{n}{numeric; the number of observations in the sample set.
May also be any object that nobs knows about;
see \code{\link[permute]{nobs}} methods.}

\item{strata}{A factor, or an object that can be coerced to a
factor via as.factor, specifying the strata for permutation.}
}
\description{
The permutation schema is defined using the `how` function.
The implementation of this function is inspired
from the VEGAN package and reproduced here to avoid an extra
dependency on an hidden vegan function.
}
\note{
Internal function do not use.
}
