% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procmod_frame.R
\name{.procmod_coerce_value}
\alias{.procmod_coerce_value}
\title{Internal function coercing the data to a matrix.}
\usage{
.procmod_coerce_value(x, nrows = 0, contrasts = NULL)
}
\arguments{
\item{x}{The data to coerce}

\item{nrows}{an interger value specifying the number of row
of the returned matrix}

\item{contrasts}{see the \code{contrasts_arg} argument
of the \code{\link[ProcMod]{procmod_frame}}
constructor.}
}
\value{
a new numeric matrix with correct size.
}
\description{
Transforme the \code{x} value into a \code{numeric matrix} of
             the correct size or into a \code{dist} object.
}
\note{
Internal function do not use.
}
\author{
Eric Coissac <eric.coissac@metabarcoding.org>

Christelle Gonindard-Melodelima <christelle.gonindard@metabarcoding.org>
}
