% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procmod_frame.R
\name{.rep_matrix}
\alias{.rep_matrix}
\title{Internal function repeating a matrix.}
\usage{
.rep_matrix(x, nrow)
}
\arguments{
\item{x}{The matrix to replicate}

\item{nrow}{an interger value specifying the number of row
of the returned matrix}
}
\value{
a new matrix with the same number of columns but with `nrow`
        rows.
}
\description{
repeats several times the rows of a matrix
             to create a new matrix with more rows. The
             final row count must be a multiple of the
             initial row count
}
\note{
Internal function do not use.
}
\author{
Eric Coissac <eric.coissac@metabarcoding.org>

Christelle Gonindard-Melodelima <christelle.gonindard@metabarcoding.org>
}
