% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covls.R
\name{.var2cor}
\alias{.var2cor}
\title{Convert a covariance matrix to a correlation matric}
\usage{
.var2cor(c)
}
\arguments{
\item{c}{a square covariance/variance matrix}
}
\value{
correlation matrix corresponding to \code{c}
}
\description{
Convert a covariance matrix to a correlation matric
}
\note{
Internal function do not use.
}
\author{
Eric Coissac

Christelle Gonindard-Melodelima
}
