% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procmod.R
\docType{package}
\name{procmod}
\alias{procmod}
\alias{procmod-package}
\title{Informative Procrustean Matrix Correlation}
\description{
Estimates corrected Procrustean correlation between
             matrices for removing overfitting effect.
}
\details{
The functions in the ProcMod package aims to estimate and to test correlation
between matrices, correcting for the spurious correlations because of the
over-fitting effect.

The ProcMod package is developed on the metabarcoding.org gitlab
(https://git.metabarcoding.org/lecasofts/ProcMod).
The gitlab of metabarcoding.org provides up-to-date information and
forums for bug reports.
}
\author{
Christelle Gonindard-Melodelima

Eric Coissac
}
